refact: ToDoList 구조 정리
