+++ src/components/createToDo.tsx
@@ -1,0 +1,35 @@
+import { useForm } from 'react-hook-form';
+import { useSetRecoilState } from 'recoil';
+import { toDoState } from '../atoms';
+
+interface IForm {
+  toDo: string;
+}
+
+function CreateToDo() {
+  const setToDos = useSetRecoilState(toDoState);
+
+  const { register, handleSubmit, setValue } = useForm<IForm>();
+
+  const handleValid = ({ toDo }: IForm) => {
+    setToDos((prev) => [
+      ...prev,
+      { id: Date.now(), text: toDo, category: 'TODO' },
+    ]);
+    setValue('toDo', '');
+  };
+
+  return (
+    <form onSubmit={handleSubmit(handleValid)}>
+      <input
+        {...register('toDo', {
+          required: 'Please write a TODO',
+        })}
+        placeholder="Write a to do"
+      />
+      <button>Add</button>
+    </form>
+  );
+}
+
+export default CreateToDo;
